requirements.txt
asgiref==3.8.1
attrs==25.3.0
coverage==7.8.2
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-gis==1.0.0 # ADDED for GeoJSON support
django-filter==24.2
drf-spectacular==0.28.0
et_xmlfile==2.0.0
inflection==0.5.1
iniconfig==2.1.0
joblib==1.5.1
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
# mysqlclient==2.2.7 # REMOVED
numpy==2.2.6
openpyxl==3.1.5
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pluggy==1.6.0
psycopg2-binary==2.9.9 # ADDED for PostgreSQL
pytest==8.3.5
pytest-django==4.11.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.1
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
matplotlib
scikit-image
seaborn